pyproject.toml
[tool.poetry]
name = "cats-and-dogs-bot"
version = "0.1.0"
description = ""
authors = ["kachus <ka4usova.e@gmail.com>"]
readme = "README.md"
packages = [{include = "cats_and_dogs_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-decouple = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
